+++ projeto/BlueHearts/src/main/java/io/github/antthluca/blue_hearts/client/ClientBlueBloodData.java
@@ -1,22 +1,22 @@
 package io.github.antthluca.blue_hearts.client;
 
 public class ClientBlueBloodData {
-    private static int playerBlueBlood = 0;
-    private static int maxBlueBlood = 0;
+    private static float playerBlueBlood = 0;
+    private static float maxBlueBlood = 0;
 
-    public static void setPlayerBlueBlood(int blueBlood) {
+    public static void setPlayerBlueBlood(float blueBlood) {
         ClientBlueBloodData.playerBlueBlood = blueBlood;
     }
 
-    public static int getPlayerBlueBlood() {
+    public static float getPlayerBlueBlood() {
         return playerBlueBlood;
     }
 
-    public static void setMaxBlueBlood(int blueBlood) {
+    public static void setMaxBlueBlood(float blueBlood) {
         ClientBlueBloodData.maxBlueBlood = blueBlood;
     }
 
-    public static int getMaxBlueBlood() {
+    public static float getMaxBlueBlood() {
         return maxBlueBlood;
     }
 }
